// apps/web/src/generated/build-version.ts
export const BUILD_VERSION: {
  label: string;
  updatedAt: string;
  commitHash: string;
  shortCommitHash: string;
  commitUrl: string | null;
} = {"label":"Updated 2026-06-17","updatedAt":"2026-06-17","commitHash":"54e5d3006c94668778efb8052e1748cd47721b30","shortCommitHash":"54e5d30","commitUrl":"https://github.com/MuzzammilRafiq/apreal/commit/54e5d3006c94668778efb8052e1748cd47721b30"};